+++ src/libYARP_run/tests/RunTest.cpp
@@ -58,7 +58,9 @@ public:
     }
 
     void stop()
-    {}
+    {
+        Thread::stop();
+    }
 
     void run() override
     {
@@ -69,15 +71,10 @@ public:
 
 TEST_CASE("run::RunTest", "[yarp::run]")
 {
-
-#if !defined(ENABLE_BROKEN_TESTS)
-    YARP_SKIP_TEST("Skipping YARP_run broken tests")
-#endif
+    NetworkBase::setLocalMode(false);
 
     SECTION("testRun")
     {
-        //this could be local or using an external nameserver, to be decided
-        Network::setLocalMode(true);
         YarpRun runner;
 
         INFO("checking yarprun");
@@ -89,20 +86,83 @@ TEST_CASE("run::RunTest", "[yarp::run]")
         argv[2]="/run";
 
         runner.start(argc, (char**)argv);
+        Time::delay(4);
+        bool ret = true;
 
-        Time::delay(3);
-        Property par;
-        par.put("name", "testModule");
+        ret = yarp::run::Run::isRunning("/run", "should_fail");
+        CHECK(!ret);
 
+        //start a process
+        std::string str1 = "test_module1";
+        Property par1;
+        par1.put("name", "testModule");
+        ret = yarp::run::Run::start("/run", par1, str1);
+        CHECK(ret);
 
-        std::string moduleTag="test_module";
-        yarp::run::Run::start("/run", par, moduleTag);
+        yarp::os::Time::delay(1);
 
-        Time::delay(1);
+        // start another process
+        std::string str2 = "test_module2";
+        Property par2;
+        par2.put("name", "testModule");
+        ret = yarp::run::Run::start("/run", par2, str2);
+        CHECK(ret);
 
-        bool isRunning=yarp::run::Run::isRunning("/run", moduleTag);
-        CHECK(isRunning); // isRunning
+        yarp::os::Time::delay(1);
+
+        // check if the processes are running
+        ret = yarp::run::Run::isRunning("/run", "test_module1");
+        CHECK(ret);
+
+        ret = yarp::run::Run::isRunning("/run", "test_module2");
+        CHECK(ret);
+
+        ret = yarp::run::Run::isRunning("/run", "nonExistingTag");
+        CHECK(!ret);
+
+        // now we have 2 processes running
+        std::vector<yarp::run::Run::processInfo> processes;
+        ret = yarp::run::Run::ps("/run",processes);
+        CHECK(processes.size()==2);
+        CHECK(ret);
+
+        //stop 1 process
+        ret = yarp::run::Run::sigterm("/run", "test_module1");
+        CHECK(ret);
+
+        yarp::os::Time::delay(1);
+
+        // now we have only 1 process running
+        std::vector<yarp::run::Run::processInfo> allprocessesclosed1;
+        ret = yarp::run::Run::ps("/run",allprocessesclosed1);
+        CHECK(allprocessesclosed1.size()==1);
+
+        // stop all processes
+        ret = yarp::run::Run::sigtermall("/run");
+        CHECK(ret);
+
+        yarp::os::Time::delay(1);
+
+        // now we have no processes running
+        std::vector<yarp::run::Run::processInfo> allprocessesclosed0;
+        ret = yarp::run::Run::ps("/run",allprocessesclosed0);
+        CHECK(allprocessesclosed0.size()==0);
+        CHECK(ret);
+
+        // check system info
+        yarp::os::SystemInfoSerializer info;
+        ret = yarp::run::Run::sysinfo("/run", info);
+        CHECK(info.memory.totalSpace!=0);
+        CHECK(ret);
+
+        // terminate the server
+        ret = yarp::run::Run::exit("/run");
+        CHECK(ret);
+
         fprintf(stderr, "done!\n");
-        Network::setLocalMode(false);
+
+        runner.stop();
     }
+
+    NetworkBase::setLocalMode(false);
 }
